Implemented the rest of the events the StreamDeck App can send our plugin
