+++ src/SamplePlugin/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "SamplePlugin": {
       "commandName": "Executable",
-      "executablePath": "cmd",
+      "executablePath": "cmd.exe",
       "commandLineArgs": "/S /C \"start \"title\" /B \"C:\\Program Files\\Elgato\\StreamDeck\\StreamDeck.exe\" \""
     }
   }
