+++ src/StreamDeckLib/Messages/Coordinates.cs
@@ -1,0 +1,15 @@
+﻿namespace StreamDeckLib.Messages
+{
+	public class Coordinates
+	{
+		public int column { get; set; }
+		public int row { get; set; }
+	}
+
+	public class Size
+	{
+		public int columns { get; set; }
+		public int rows { get; set; }
+	}
+
+}
